# apidocs/swagger.json
{
  "consumes": [
    "application/json",
    "text/plain"
  ],
  "produces": [
    "application/json",
    "text/plain"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes the Supergiant Control API.",
    "title": "Supergiant Control API.",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "2.1.2"
  },
  "basePath": "/",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a new user.",
        "operationId": "createUser",
        "parameters": [
          {
            "name": "Req",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/auth": {
      "post": {
        "tags": [
          "signin"
        ],
        "summary": "Authenticate a user.",
        "operationId": "auth",
        "parameters": [
          {
            "name": "Req",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AuthRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/authResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/coldstart": {
      "get": {
        "tags": [
          "signin"
        ],
        "summary": "Tells if any users are registered.",
        "operationId": "coldstart",
        "responses": {
          "200": {
            "$ref": "#/responses/coldstartResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/root": {
      "post": {
        "tags": [
          "signin"
        ],
        "summary": "Register the root user.",
        "operationId": "registerRoot",
        "parameters": [
          {
            "name": "Req",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/helm/repositories": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "List helm repositories.",
        "operationId": "listRepos",
        "responses": {
          "200": {
            "$ref": "#/responses/listReposResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "repositories"
        ],
        "summary": "Create a helm repository.",
        "operationId": "createRepo",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Entry"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/repositoryResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/helm/repositories/{repoName}": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Get a helm repository.",
        "operationId": "getRepo",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "RepoName",
            "name": "repoName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/repositoryResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "repositories"
        ],
        "summary": "Delete a helm repository.",
        "operationId": "deleteRepo",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "RepoName",
            "name": "repoName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/repositoryResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "patch": {
        "tags": [
          "repositories"
        ],
        "summary": "Update a helm repository.",
        "operationId": "updateRepo",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "RepoName",
            "name": "repoName",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Entry"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/repositoryResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/helm/repositories/{repoName}/charts": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Get charts for the given helm repository.",
        "operationId": "listCharts",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "RepoName",
            "name": "repoName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listChartsResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/helm/repositories/{repoName}/charts/{chartName}": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Get a chart from the given helm repository.",
        "operationId": "getChart",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "RepoName",
            "name": "repoName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "chartName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/chartResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
    "/v1/api/provision": {
      "post": {
        "tags": [
          "kubes"
        ],
        "summary": "Provision a kubernetes cluster.",
        "operationId": "provision",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ProvisionRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "$ref": "#/responses/provisionResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Addresses": {
      "type": "object",
      "title": "Addresses uses cidr to define an ip list.",
      "properties": {
        "cidr": {
          "type": "string",
          "x-go-name": "CIDR"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "AuthRequest": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/user"
    },
...
    "ChartData": {
      "type": "object",
      "title": "ChartData is a simplified representation of the helm chart.",
      "properties": {
        "metadata": {
          "$ref": "#/definitions/Metadata"
        },
        "readme": {
          "type": "string",
          "x-go-name": "Readme"
        },
        "values": {
          "type": "string",
          "x-go-name": "Values"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/model"
    },
    "ChartInfo": {
      "description": "ChartInfo contains a list of available chart versions and description/icon\ninfo for the latest one.",
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "icon": {
          "type": "string",
          "x-go-name": "Icon"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "repo": {
          "type": "string",
          "x-go-name": "Repo"
        },
        "versions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ChartVersion"
          },
          "x-go-name": "Versions"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/model"
    },
...
    "CloudSpecificSettings": {
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
    "ColdstartResponse": {
      "type": "object",
      "properties": {
        "isColdStart": {
          "type": "boolean",
          "x-go-name": "IsColdStart"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/user"
    },
...
    "Entry": {
      "description": "Entry represents a collection of parameters for chart repository",
      "type": "object",
      "properties": {
        "caFile": {
          "type": "string",
          "x-go-name": "CAFile"
        },
        "cache": {
          "type": "string",
          "x-go-name": "Cache"
        },
        "certFile": {
          "type": "string",
          "x-go-name": "CertFile"
        },
        "keyFile": {
          "type": "string",
          "x-go-name": "KeyFile"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "url": {
          "type": "string",
          "x-go-name": "URL"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "k8s.io/helm/pkg/repo"
    },
    "ErrorCode": {
      "type": "integer",
      "format": "int64",
      "x-go-package": "github.com/supergiant/control/pkg/sgerrors"
    },
...
    "Message": {
      "description": "Message is used in response body to display separate messages to user and developers",
      "type": "object",
      "properties": {
        "devMessage": {
          "description": "DevMessage should be used to display message to developer on API/console",
          "type": "string",
          "x-go-name": "DevMessage"
        },
        "errorCode": {
          "$ref": "#/definitions/ErrorCode"
        },
        "moreInfo": {
          "description": "MoreInfo should be a link to supergiant documentation to display common problems",
          "type": "string",
          "x-go-name": "MoreInfo"
        },
        "userMessage": {
          "description": "UserMessage should be used to display message to user on UI/CLI",
          "type": "string",
          "x-go-name": "UserMessage"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/message"
    },
...
    "Profile": {
      "type": "object",
      "properties": {
        "K8SVersion": {
          "type": "string"
        },
        "addons": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Addons"
        },
        "arch": {
          "type": "string",
          "x-go-name": "Arch"
        },
        "cidr": {
          "type": "string",
          "x-go-name": "CIDR"
        },
        "cloudSpecificSettings": {
          "$ref": "#/definitions/CloudSpecificSettings"
        },
        "dockerVersion": {
          "type": "string",
          "x-go-name": "DockerVersion"
        },
        "exposedAddresses": {
          "description": "ExposedAddresses is a list of cidr/port pairs that will be exposes\nby cloud provider security groups.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Addresses"
          },
          "x-go-name": "ExposedAddresses"
        },
        "flannelVersion": {
          "type": "string",
          "x-go-name": "FlannelVersion"
        },
        "helmVersion": {
          "type": "string",
          "x-go-name": "HelmVersion"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "k8sApiPort": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "K8SAPIPort"
        },
        "k8sServicesCIDR": {
          "type": "string",
          "x-go-name": "K8SServicesCIDR"
        },
        "masterProfiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NodeProfile"
          },
          "x-go-name": "MasterProfiles"
        },
        "networkProvider": {
          "type": "string",
          "x-go-name": "NetworkProvider"
        },
        "networkType": {
          "type": "string",
          "x-go-name": "NetworkType"
        },
        "nodesProfiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NodeProfile"
          },
          "x-go-name": "NodesProfiles"
        },
        "operatingSystem": {
          "type": "string",
          "x-go-name": "OperatingSystem"
        },
        "password": {
          "description": "DEPRECATED: should be a part of the static auth",
          "type": "string",
          "x-go-name": "Password"
        },
        "provider": {
          "$ref": "#/definitions/Name"
        },
        "publicKey": {
          "type": "string",
          "x-go-name": "PublicKey"
        },
        "rbacEnabled": {
          "type": "boolean",
          "x-go-name": "RBACEnabled"
        },
        "region": {
          "type": "string",
          "x-go-name": "Region"
        },
        "staticAuth": {
          "$ref": "#/definitions/StaticAuth"
        },
        "subnets": {
          "description": "This field is AWS specific, mapping AZ -\u003e subnet",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Subnets"
        },
        "ubuntuVersion": {
          "type": "string",
          "x-go-name": "UbuntuVersion"
        },
        "user": {
          "description": "TODO: get rid of its usage\nDEPRECATED: should be a part of the static auth",
          "type": "string",
          "x-go-name": "User"
        },
        "zone": {
          "type": "string",
          "x-go-name": "Zone"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
    "ProvisionRequest": {
      "type": "object",
      "properties": {
        "cloudAccountName": {
          "type": "string",
          "x-go-name": "CloudAccountName"
        },
        "clusterName": {
          "type": "string",
          "x-go-name": "ClusterName"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/provisioner"
    },
    "ProvisionResponse": {
      "type": "object",
      "properties": {
        "clusterId": {
          "type": "string",
          "x-go-name": "ClusterID"
        },
        "tasks": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "x-go-name": "Tasks"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/provisioner"
    },
...
    "ReleaseInfo": {
      "type": "object",
      "title": "ReleaseInfo is a simplified representations of the helm release.",
      "properties": {
        "chart": {
          "type": "string",
          "x-go-name": "Chart"
        },
        "chartVersion": {
          "type": "string",
          "x-go-name": "ChartVersion"
        },
        "createdAt": {
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "lastDeployed": {
          "type": "string",
          "x-go-name": "LastDeployed"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "version": {
          "type": "integer",
          "format": "int32",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/model"
    },
...
    "User": {
      "description": "User is the representation of supergiant user",
      "type": "object",
      "properties": {
        "encrypted_password": {
          "type": "string",
          "format": "string",
          "x-go-name": "EncryptedPassword"
        },
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/user"
    },
...
  },
  "responses": {
...
    "authResponse": {
      "description": "authResponse contains an auth header.",
      "headers": {
        "Access-Control-Expose-Headers": {
          "type": "string",
          "description": "AccessControlExposeHeaders header"
        },
        "Authorization": {
          "type": "string",
          "description": "Authorization header"
        }
      }
    },
    "chartResponse": {
      "description": "chartResponse contains representations of a helm chart.",
      "schema": {
        "$ref": "#/definitions/ChartData"
      }
    },
...
    "coldstartResponse": {
      "description": "coldstartResponse tells if there is any user exists.",
      "schema": {
        "$ref": "#/definitions/ColdstartResponse"
      }
    },
...
    "emptyResponse": {
      "description": "emptyResponse represents an empty http response."
    },
    "errorResponse": {
      "description": "errorResponse represents an error response.",
      "schema": {
        "$ref": "#/definitions/Message"
      }
    },
...
    "listChartsResponse": {
      "description": "listChartsResponse contains a list of helm charts.",
      "schema": {
        "$ref": "#/definitions/ChartInfo"
      }
    },
...
    "listReposResponse": {
      "description": "listReposResponse contains a list of helm repos.",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/ReleaseInfo"
        }
      }
    },
...
    "provisionResponse": {
      "description": "provisionResponse contains a cluster provisioning metadata.",
      "schema": {
        "$ref": "#/definitions/ProvisionResponse"
      }
    },
...
    "repositoryResponse": {
      "description": "repositoryResponse contains representations of a helm repository.",
      "schema": {
        "$ref": "#/definitions/ReleaseInfo"
      }
    },
...
  },
  "securityDefinitions": {
    "AuthorizationHeaderToken": {
      "description": "This scheme uses the header \"Authorization: Bearer {JWT token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}